support building test but not updating the related repos(e.g. ceresmeta, tsbs, dist-query-testing...).

+++ integration_tests/build_meta.sh
@@ -5,11 +5,15 @@ set -e
 SRC=/tmp/ceresmeta-src
 TARGET=$(pwd)/ceresmeta
 
-if [ -d $SRC ]; then
+if [[ -d $SRC && $UPDATE_REPOS_TO_LATEST == 'true' ]]; then
   echo "Remove old meta..."
   rm -rf $SRC
 fi
 
-git clone --depth 1 https://github.com/ceresdb/ceresmeta.git ${SRC}
+if [[ ! -d $SRC ]]; then
+  echo "Pull meta repo..."
+  git clone --depth 1 https://github.com/ceresdb/ceresmeta.git ${SRC}
+fi
+
 cd ${SRC}
 go build -o ${TARGET}/ceresmeta ./cmd/meta/...
